+++ src/nl/logiconline/neptune/system/geom/Rectangle.java
@@ -16,7 +16,7 @@ public class Rectangle extends Shape {
 	private boolean fill;
 	private int width, height;
 
-	public Rectangle(Gfx g, int x, int y, int width, int height, boolean fill) {
+	public Rectangle(Gfx g, double x, double y, int width, int height, boolean fill) {
 		super(g, x, y);
 		this.width = width;
 		this.height = height;
@@ -26,29 +26,29 @@ public class Rectangle extends Shape {
 	@Override
 	public void draw() throws NeptuneException {
 		if (this.fill) {
-			for (int x = super.x; x < (super.x + this.width); x++) {
-				for (int y = super.y; y < (super.y + this.height); y++) {
+			for (double x = super.x; x < (super.x + this.width); x++) {
+				for (double y = super.y; y < (super.y + this.height); y++) {
 					super.g.setPixel(x, y);
 				}
 			}
 		} else {
 			//Draw top line
-			for (int tx = this.x; tx <= (this.x + this.width); tx++) {
+			for (double tx = this.x; tx <= (this.x + this.width); tx++) {
 				this.g.setPixel(tx, this.y);
 			}
 
 			//Draw bottom line
-			for (int tx = this.x; tx <= (this.x + this.width); tx++) {
+			for (double tx = this.x; tx <= (this.x + this.width); tx++) {
 				this.g.setPixel(tx, (this.y + this.height));
 			}
 
 			//Draw Left line
-			for (int ty = this.y; ty <= (this.y + this.height); ty++) {
+			for (double ty = this.y; ty <= (this.y + this.height); ty++) {
 				this.g.setPixel(this.x, ty);
 			}
 
 			//Draw right line
-			for (int ty = this.y; ty <= ((this.y + this.height)); ty++) {
+			for (double ty = this.y; ty <= ((this.y + this.height)); ty++) {
 				this.g.setPixel(this.x + this.width, ty);
 			}
 		}
